Refactors plane creaction
